requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7
django-discover-runner==1.0
django-debug-toolbar==0.11
psycopg2
fabric